# Avoid String.split when parsing header to reduce some garbage generat…